refactor(#120): extraheer BerichtPipeline uit EmailProcessorFunction
ValideerDagDatum, VerwerkMetPlannerAsync en BouwTemplateAntwoord zijn
verplaatst naar SportlinkFunction.Processing.BerichtPipeline. Zowel de
live email-verwerker als de dry-run tester roepen nu dezelfde pipeline
aan zonder directe koppeling via EmailProcessorFunction.
