+++ FunctionApp/Email/EmailProcessorFunction.cs
@@ -4,7 +4,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Microsoft.Graph;
 using Newtonsoft.Json;
-using SportlinkFunction.Planner;
+using SportlinkFunction.Processing;
 
 namespace SportlinkFunction.Email;
 
@@ -154,8 +154,7 @@ public class EmailProcessorFunction
         var classificatie = await aiService.ClassificeerBerichtAsync(
             email.Body, email.Onderwerp, email.Afzender);
 
-        // Valideer dag-datum: extraheer uit onderwerp (prioriteit) en body, corrigeer AI-fouten
-        ValideerDagDatum(classificatie, email.Body, email.Onderwerp);
+        BerichtPipeline.ValideerDagDatum(classificatie, email.Body, email.Onderwerp);
 
         var classificatieJson = JsonConvert.SerializeObject(classificatie);
         await UpdateStatusAsync(verwerkingId, EmailStatus.Geclassificeerd, classificatieJson);
@@ -175,11 +174,11 @@ public class EmailProcessorFunction
             return;
         }
 
-        var plannerResponseJson = await VerwerkMetPlannerAsync(classificatie, email, log);
+        var plannerResponseJson = await BerichtPipeline.VerwerkMetPlannerAsync(classificatie, email, log);
         await UpdatePlannerResponseAsync(verwerkingId, plannerResponseJson);
         await UpdateStatusAsync(verwerkingId, EmailStatus.Verwerkt, null);
 
-        var (onderwerp, antwoordBody) = BouwTemplateAntwoord(
+        var (onderwerp, antwoordBody) = BerichtPipeline.BouwTemplateAntwoord(
             classificatie, plannerResponseJson, email);
 
         var reviewMode = Environment.GetEnvironmentVariable("EmailReviewMode");
@@ -193,426 +192,6 @@ public class EmailProcessorFunction
 
         log.LogInformation("Email {Id} volledig verwerkt, antwoord verstuurd naar {Ontvanger}",
             verwerkingId, ontvanger);
-    }
-
-    /// <summary>
-    /// Bouwt het antwoord via templates op basis van het classificatietype en PlannerService response.
-    /// </summary>
-    internal static (string onderwerp, string body) BouwTemplateAntwoord(
-        BerichtClassificatie classificatie,
-        string plannerResponseJson,
-        InkomendBericht email)
-    {
-        switch (classificatie.Type)
-        {
-            case VerzoekType.BeschikbaarheidCheck:
-                var jobj = Newtonsoft.Json.Linq.JObject.Parse(plannerResponseJson);
-
-                if (jobj["wedstrijdAlIngepland"]?.ToObject<bool>() == true)
-                {
-                    var ingeplandWedstrijd = jobj["wedstrijd"]?.ToObject<ZoekWedstrijdResponse>();
-                    return BerichtResponseGenerator.BouwWedstrijdAlIngeplandAntwoord(
-                        ingeplandWedstrijd, classificatie, email);
-                }
-
-                if (jobj["teamOnbekend"]?.ToObject<bool>() == true)
-                {
-                    var onbekendeTegenstander = jobj["tegenstander"]?.ToString()
-                        ?? classificatie.Tegenstander ?? "";
-                    return BerichtResponseGenerator.BouwTeamOnbekendAntwoord(
-                        onbekendeTegenstander, classificatie, email);
-                }
-
-                if (jobj["multiDatum"]?.ToObject<bool>() == true)
-                {
-                    var resultaten = new List<(string datum, CheckAvailabilityResponse response)>();
-                    foreach (var item in jobj["resultaten"]!)
-                    {
-                        var datum = item["datum"]?.ToString() ?? "";
-                        var resp = item["response"]?.ToObject<CheckAvailabilityResponse>() ?? new CheckAvailabilityResponse();
-                        resultaten.Add((datum, resp));
-                    }
-                    return BerichtResponseGenerator.BouwMultiDatumBeschikbaarheidAntwoord(
-                        resultaten, classificatie, email);
-                }
-                var checkResponse = JsonConvert.DeserializeObject<CheckAvailabilityResponse>(plannerResponseJson);
-                return BerichtResponseGenerator.BouwBeschikbaarheidAntwoord(
-                    checkResponse ?? new CheckAvailabilityResponse(), classificatie, email);
-
-            case VerzoekType.HerplanVerzoek:
-                var herplanData = Newtonsoft.Json.Linq.JObject.Parse(plannerResponseJson);
-                var wedstrijd = herplanData["wedstrijd"]?.ToObject<ZoekWedstrijdResponse>();
-
-                // #67 — Te laat verzonden herplanverzoek
-                if (herplanData["herplanTeLaat"]?.ToObject<bool>() == true)
-                {
-                    var teLaatWedstrijd = herplanData["wedstrijd"]?.ToObject<ZoekWedstrijdResponse>();
-                    var deadlineDagen = herplanData["deadlineDagen"]?.ToObject<int>() ?? 8;
-                    var dagenTot = herplanData["dagenTotWedstrijd"]?.ToObject<int>() ?? 0;
-                    return BerichtResponseGenerator.BouwHerplanTeLaatAntwoord(teLaatWedstrijd, deadlineDagen, dagenTot, classificatie, email);
-                }
-
-                // Gewenste datum modus: beschikbaarheid op nieuwe datum
-                if (herplanData["gewensteDatum"] != null && herplanData["beschikbaarheid"] != null)
-                {
-                    var gewensteDatum = herplanData["gewensteDatum"]?.ToString();
-                    var beschikbaarheid = herplanData["beschikbaarheid"]?.ToObject<CheckAvailabilityResponse>();
-                    return BerichtResponseGenerator.BouwHerplanGewensteDatumAntwoord(
-                        wedstrijd, gewensteDatum, beschikbaarheid, classificatie, email);
-                }
-
-                // Alternatieven op huidige dag
-                var herplanOpties = herplanData["herplanOpties"]?.ToObject<HerplanCheckResponse>();
-                return BerichtResponseGenerator.BouwHerplanAntwoord(
-                    wedstrijd, herplanOpties, classificatie, email);
-
-            case VerzoekType.Bevestiging:
-                return BerichtResponseGenerator.BouwBevestigingAntwoord(email, classificatie);
-
-            default:
-                return BerichtResponseGenerator.BouwBuitenScopeAntwoord(email);
-        }
-    }
-
-    /// <summary>
-    /// Extraheert datums uit onderwerp en body, en corrigeert de AI-classificatie.
-    /// Prioriteit: expliciete datum in onderwerp > expliciete datum in body > AI datum + dag-validatie.
-    /// </summary>
-    internal static void ValideerDagDatum(BerichtClassificatie classificatie, string emailBody, string onderwerp)
-    {
-        // Stap 1: Zoek expliciete datum in onderwerp (bijv. "18-4-2026", "18-04-2026", "9 mei")
-        var onderwerpDatum = ExtractExpliciteDatum(onderwerp);
-        if (onderwerpDatum.HasValue)
-        {
-            classificatie.Datum = onderwerpDatum.Value.ToString("yyyy-MM-dd");
-            return;
-        }
-
-        // Stap 2: Zoek expliciete datum in body (eerste dd-mm-yyyy patroon)
-        var bodyDatum = ExtractExpliciteDatum(emailBody);
-        if (bodyDatum.HasValue && string.IsNullOrEmpty(classificatie.Datum))
-        {
-            classificatie.Datum = bodyDatum.Value.ToString("yyyy-MM-dd");
-            return;
-        }
-
-        // Stap 3: Dag-naam validatie als fallback
-        if (string.IsNullOrEmpty(classificatie.Datum)) return;
-        if (!DateOnly.TryParse(classificatie.Datum, out var datum)) return;
-
-        var tekst = (onderwerp + " " + emailBody).ToLowerInvariant();
-        var dagNamen = new (string naam, DayOfWeek dag)[]
-        {
-            ("maandag", DayOfWeek.Monday), ("dinsdag", DayOfWeek.Tuesday),
-            ("woensdag", DayOfWeek.Wednesday), ("donderdag", DayOfWeek.Thursday),
-            ("vrijdag", DayOfWeek.Friday), ("zaterdag", DayOfWeek.Saturday),
-            ("zondag", DayOfWeek.Sunday)
-        };
-
-        foreach (var (naam, dag) in dagNamen)
-        {
-            if (!tekst.Contains(naam)) continue;
-            if (datum.DayOfWeek == dag) return;
-
-            for (int offset = 1; offset <= 7; offset++)
-            {
-                if (datum.AddDays(-offset).DayOfWeek == dag)
-                    { classificatie.Datum = datum.AddDays(-offset).ToString("yyyy-MM-dd"); return; }
-                if (datum.AddDays(offset).DayOfWeek == dag)
-                    { classificatie.Datum = datum.AddDays(offset).ToString("yyyy-MM-dd"); return; }
-            }
-            return;
-        }
-    }
-
-    /// <summary>
-    /// Extraheert een expliciete datum uit tekst. Herkent patronen als "18-4-2026", "18-04-2026", "9 mei 2026", "9 mei".
-    /// </summary>
-    private static DateOnly? ExtractExpliciteDatum(string tekst)
-    {
-        if (string.IsNullOrWhiteSpace(tekst)) return null;
-
-        // Patroon: dd-mm-yyyy of d-m-yyyy
-        var numericMatch = System.Text.RegularExpressions.Regex.Match(tekst, @"(\d{1,2})-(\d{1,2})-(\d{4})");
-        if (numericMatch.Success)
-        {
-            if (int.TryParse(numericMatch.Groups[1].Value, out var dag) &&
-                int.TryParse(numericMatch.Groups[2].Value, out var maand) &&
-                int.TryParse(numericMatch.Groups[3].Value, out var jaar))
-            {
-                try { return new DateOnly(jaar, maand, dag); } catch { }
-            }
-        }
-
-        // Patroon: "d maand" of "d maand yyyy" (bijv. "9 mei", "25 april 2026")
-        var maandNamen = new Dictionary<string, int>(StringComparer.OrdinalIgnoreCase)
-        {
-            ["januari"] = 1, ["februari"] = 2, ["maart"] = 3, ["april"] = 4,
-            ["mei"] = 5, ["juni"] = 6, ["juli"] = 7, ["augustus"] = 8,
-            ["september"] = 9, ["oktober"] = 10, ["november"] = 11, ["december"] = 12
-        };
-
-        var tekstLower = tekst.ToLowerInvariant();
-        foreach (var (naam, maandNr) in maandNamen)
-        {
-            var maandMatch = System.Text.RegularExpressions.Regex.Match(tekstLower, $@"(\d{{1,2}})\s+{naam}(?:\s+(\d{{4}}))?");
-            if (maandMatch.Success && int.TryParse(maandMatch.Groups[1].Value, out var d))
-            {
-                var j = maandMatch.Groups[2].Success && int.TryParse(maandMatch.Groups[2].Value, out var jj)
-                    ? jj : DateTime.Now.Year;
-                try { return new DateOnly(j, maandNr, d); } catch { }
-            }
-        }
-
-        return null;
-    }
-
-    /// <summary>
-    /// Extraheert het eigen team uit een wedstrijdnaam als "CLUBCODE JO16-2 - Tegenstander JO16-4".
-    /// </summary>
-    private static string? ExtractVrcTeamUitWedstrijd(string wedstrijd, string tegenstander)
-    {
-        var clubPrefix = (SystemUtilities.AppSettings.GetSetting("clubCode") ?? "") + " ";
-        var parts = wedstrijd.Split(" - ", 2, StringSplitOptions.TrimEntries);
-        foreach (var part in parts)
-            if (!string.IsNullOrWhiteSpace(clubPrefix.Trim())
-                && part.StartsWith(clubPrefix, StringComparison.OrdinalIgnoreCase)
-                && !part.Contains(tegenstander, StringComparison.OrdinalIgnoreCase))
-                return part;
-        foreach (var part in parts)
-            if (!part.Contains(tegenstander, StringComparison.OrdinalIgnoreCase))
-                return part;
-        return null;
-    }
-
-    /// <summary>
-    /// Normaliseert teamnaam: voegt clubprefix toe, normaliseert JO/MO/O-prefix, vervangt / door -.
-    /// </summary>
-    private static string? NormaliseerTeamNaam(string? teamNaam)
-    {
-        if (string.IsNullOrWhiteSpace(teamNaam)) return teamNaam;
-        var t = teamNaam.Trim();
-
-        // Slash tussen cijfers → streep (bijv. "JO17/3" → "JO17-3", "17/1" → "17-1").
-        // Trainers schrijven vaak "jo17/3" terwijl Sportlink "JO17-3" gebruikt; zonder
-        // deze normalisatie faalt de LIKE-zoekopdracht in FindMatchAsync.
-        t = System.Text.RegularExpressions.Regex.Replace(t, @"(\d)\s*/\s*(\d)", "$1-$2");
-
-        // "Onder 13-1" → "JO13-1"
-        if (t.StartsWith("Onder ", StringComparison.OrdinalIgnoreCase))
-            t = "JO" + t[6..].Trim();
-
-        // "O13-1" → "JO13-1" (maar niet "MO13-1")
-        if (System.Text.RegularExpressions.Regex.IsMatch(t, @"^O\d", System.Text.RegularExpressions.RegexOptions.IgnoreCase)
-            && !t.StartsWith("MO", StringComparison.OrdinalIgnoreCase))
-            t = "J" + t.ToUpper();
-
-        // Clubprefix toevoegen als de naam er uitziet als een eigen team
-        // (start met JO/MO/VR/JM/ZO + cijfer, of is puur cijfer/letters zonder clubnaam).
-        // Namen met spatie erin zijn meestal externe clubs (bijv. "VOP JO14-1", "Hooglanderveen JO16-4").
-        bool looksLikeVrcTeam = System.Text.RegularExpressions.Regex.IsMatch(t, @"^(JO|MO|VR|JM|ZO)\d", System.Text.RegularExpressions.RegexOptions.IgnoreCase)
-                              || !t.Contains(' ');
-        var clubCode = SystemUtilities.AppSettings.GetSetting("clubCode");
-        if (!string.IsNullOrWhiteSpace(clubCode))
-        {
-            var clubPrefix = clubCode + " ";
-            if (looksLikeVrcTeam && !t.StartsWith(clubPrefix, StringComparison.OrdinalIgnoreCase))
-                t = clubPrefix + t;
-        }
-
-        return t;
-    }
-
-    /// <summary>
-    /// Normaliseert leeftijdscategorie: O13 → JO13, Onder 11 → JO11, etc.
-    /// </summary>
-    private static string? NormaliseerLeeftijdsCategorie(string? categorie)
-    {
-        if (string.IsNullOrWhiteSpace(categorie)) return categorie;
-        var c = categorie.Trim();
-        // "Onder 13" → "JO13"
-        if (c.StartsWith("Onder ", StringComparison.OrdinalIgnoreCase))
-            c = "JO" + c[6..].Trim();
-        // "O13" → "JO13" (maar niet "MO13")
-        if (System.Text.RegularExpressions.Regex.IsMatch(c, @"^O\d", System.Text.RegularExpressions.RegexOptions.IgnoreCase)
-            && !c.StartsWith("MO", StringComparison.OrdinalIgnoreCase))
-            c = "J" + c.ToUpper();
-        return c;
-    }
-
-    /// <summary>
-    /// Bepaalt of een herplanverzoek vervroeging of verlating betreft, op basis van trefwoorden
-    /// in onderwerp en body. Geeft null terug als beide of geen van beide voorkomen — dan vallen
-    /// we terug op het standaardgedrag (alle slots, gesorteerd op nabijheid).
-    /// </summary>
-    private static string? DetecteerRichting(string onderwerp, string body)
-    {
-        var tekst = ((onderwerp ?? "") + " " + (body ?? "")).ToLowerInvariant();
-        bool vervroegen = tekst.Contains("vervroeg") || tekst.Contains("eerder")
-                       || tekst.Contains("naar voren");
-        bool verlaten = tekst.Contains("verlaat") || tekst.Contains("verlat")
-                     || tekst.Contains(" later") || tekst.Contains("naar achter");
-        if (vervroegen && !verlaten) return "vervroegen";
-        if (verlaten && !vervroegen) return "verlaten";
-        return null;
-    }
-
-    /// <summary>
-    /// Vertaalt de AI-classificatie naar de juiste PlannerService-aanroep.
-    /// </summary>
-    internal static async Task<string> VerwerkMetPlannerAsync(
-        BerichtClassificatie classificatie, InkomendBericht email, ILogger log)
-    {
-        classificatie.LeeftijdsCategorie = NormaliseerLeeftijdsCategorie(classificatie.LeeftijdsCategorie);
-
-        // Bepaal welk veld het eigen team is (kan in TeamNaam of Tegenstander staan)
-        var team = classificatie.TeamNaam ?? "";
-        var tegenstander = classificatie.Tegenstander ?? "";
-
-        // Heuristiek: naam zonder clubnaam erin is waarschijnlijk het eigen team
-        // Externe teams bevatten vaak hun clubnaam (bijv. "Woudenberg MO15-1", "Hooglanderveen JO16-4")
-        // Swap alleen als BEIDE velden gevuld zijn en de rollen duidelijk omgedraaid zijn
-        if (!string.IsNullOrWhiteSpace(team) && !string.IsNullOrWhiteSpace(tegenstander))
-        {
-            var cc = SystemUtilities.AppSettings.GetSetting("clubCode") ?? "";
-            bool teamIsEigenClub = !team.Contains(' ')
-                || (!string.IsNullOrWhiteSpace(cc) && team.StartsWith(cc, StringComparison.OrdinalIgnoreCase));
-            bool tegenstanderIsEigenClub = !tegenstander.Contains(' ')
-                || (!string.IsNullOrWhiteSpace(cc) && tegenstander.StartsWith(cc, StringComparison.OrdinalIgnoreCase));
-
-            if (!teamIsEigenClub && tegenstanderIsEigenClub)
-            {
-                // Tegenstander-veld bevat het eigen team → swap
-                classificatie.TeamNaam = tegenstander;
-                classificatie.Tegenstander = team;
-            }
-        }
-
-        classificatie.TeamNaam = NormaliseerTeamNaam(classificatie.TeamNaam);
-
-        switch (classificatie.Type)
-        {
-            case VerzoekType.BeschikbaarheidCheck:
-                var alleDatums = classificatie.GetAlleDatums();
-
-                // Tegenstander-check: als een extern team vraagt, controleer eerst of de wedstrijd
-                // al in het programma staat vóórdat we beschikbaarheid tonen.
-                // Alleen bij single-datum requests (multi-datum is altijd een open vraag).
-                var cc2 = SystemUtilities.AppSettings.GetSetting("clubCode") ?? "";
-                bool heeftExterneTegenstander = !string.IsNullOrWhiteSpace(classificatie.Tegenstander)
-                    && (string.IsNullOrWhiteSpace(cc2)
-                        || !classificatie.Tegenstander.StartsWith(cc2, StringComparison.OrdinalIgnoreCase));
-                bool heeftOnbekendVrcTeam = string.IsNullOrWhiteSpace(classificatie.TeamNaam)
-                    || string.IsNullOrWhiteSpace(cc2)
-                    || !classificatie.TeamNaam.StartsWith(cc2, StringComparison.OrdinalIgnoreCase);
-
-                if (heeftExterneTegenstander && heeftOnbekendVrcTeam && alleDatums.Count == 1
-                    && DateOnly.TryParse(alleDatums[0], out var opponentCheckDatum))
-                {
-                    var wedstrijdOpDatum = await PlannerDataAccess.FindMatchByOpponentAsync(
-                        classificatie.Tegenstander!, opponentCheckDatum);
-                    if (wedstrijdOpDatum != null)
-                        return JsonConvert.SerializeObject(new { wedstrijdAlIngepland = true, wedstrijd = wedstrijdOpDatum });
-
-                    var wedstrijdAndereDatum = await PlannerDataAccess.FindMatchByOpponentAsync(
-                        classificatie.Tegenstander!, null);
-                    if (wedstrijdAndereDatum == null)
-                        return JsonConvert.SerializeObject(new { teamOnbekend = true, tegenstander = classificatie.Tegenstander });
-
-                    var vrcTeam = ExtractVrcTeamUitWedstrijd(wedstrijdAndereDatum.Wedstrijd, classificatie.Tegenstander!);
-                    if (vrcTeam != null)
-                        classificatie.TeamNaam = NormaliseerTeamNaam(vrcTeam);
-                }
-
-                if (alleDatums.Count > 1)
-                {
-                    var multiResults = new List<object>();
-                    foreach (var datum in alleDatums)
-                    {
-                        var req = new CheckAvailabilityRequest
-                        {
-                            Datum = datum,
-                            AanvangsTijd = classificatie.AanvangsTijd,
-                            LeeftijdsCategorie = classificatie.LeeftijdsCategorie,
-                            TeamNaam = classificatie.TeamNaam,
-                            Tegenstander = classificatie.Tegenstander
-                        };
-                        var resp = await PlannerService.CheckAvailabilityAsync(req, log);
-                        multiResults.Add(new { datum, response = resp });
-                    }
-                    return JsonConvert.SerializeObject(new { multiDatum = true, resultaten = multiResults });
-                }
-                var checkRequest = new CheckAvailabilityRequest
-                {
-                    Datum = classificatie.Datum ?? "",
-                    AanvangsTijd = classificatie.AanvangsTijd,
-                    LeeftijdsCategorie = classificatie.LeeftijdsCategorie,
-                    TeamNaam = classificatie.TeamNaam,
-                    Tegenstander = classificatie.Tegenstander
-                };
-                var checkResponse = await PlannerService.CheckAvailabilityAsync(checkRequest, log);
-                return JsonConvert.SerializeObject(checkResponse);
-
-            case VerzoekType.HerplanVerzoek:
-                if (!string.IsNullOrEmpty(classificatie.TeamNaam) && !string.IsNullOrEmpty(classificatie.Datum))
-                {
-                    if (DateOnly.TryParse(classificatie.Datum, out var datum))
-                    {
-                        var wedstrijd = await PlannerDataAccess.FindMatchAsync(classificatie.TeamNaam, datum);
-                        if (wedstrijd != null)
-                        {
-                            var deadlineDagen = int.TryParse(SystemUtilities.AppSettings.GetSetting("herplanDeadlineDagen"), out var dd) ? dd : 8;
-                            if (DateOnly.TryParse(wedstrijd.Datum, out var wedstrijdDatum))
-                            {
-                                var dagenTotWedstrijd = (wedstrijdDatum.ToDateTime(TimeOnly.MinValue) - DateTime.Today).TotalDays;
-                                if (dagenTotWedstrijd < deadlineDagen)
-                                {
-                                    return JsonConvert.SerializeObject(new
-                                    {
-                                        herplanTeLaat = true,
-                                        wedstrijd,
-                                        deadlineDagen,
-                                        dagenTotWedstrijd = (int)dagenTotWedstrijd
-                                    });
-                                }
-                            }
-
-                            if (!string.IsNullOrEmpty(classificatie.GewensteDatum))
-                            {
-                                var gewenstRequest = new CheckAvailabilityRequest
-                                {
-                                    Datum = classificatie.GewensteDatum,
-                                    LeeftijdsCategorie = classificatie.LeeftijdsCategorie,
-                                    TeamNaam = classificatie.TeamNaam
-                                };
-                                var beschikbaarheid = await PlannerService.CheckAvailabilityAsync(gewenstRequest, log);
-                                return JsonConvert.SerializeObject(new { wedstrijd, gewensteDatum = classificatie.GewensteDatum, beschikbaarheid });
-                            }
-
-                            // Geen gewenste datum → check alternatieven op huidige dag.
-                            // Richtingdetectie op trefwoorden in onderwerp+body: trainers schrijven
-                            // expliciet "vervroegen"/"eerder" of "later"/"verlaten"; bij beide of
-                            // geen van beide blijft Richting null (default-gedrag, voor- én nakijken).
-                            var herplanRequest = new HerplanCheckRequest
-                            {
-                                Wedstrijdcode = wedstrijd.Wedstrijdcode,
-                                VoorkeurTijd = classificatie.AanvangsTijd,
-                                Richting = DetecteerRichting(email.Onderwerp, email.Body)
-                            };
-                            var herplanResponse = await PlannerService.CheckRescheduleAvailabilityAsync(herplanRequest, log);
-                            return JsonConvert.SerializeObject(new { wedstrijd, herplanOpties = herplanResponse });
-                        }
-                        return JsonConvert.SerializeObject(new { gevonden = false, reden = $"Geen wedstrijd gevonden voor {classificatie.TeamNaam} op {classificatie.Datum}" });
-                    }
-                }
-                return JsonConvert.SerializeObject(new { error = "Onvoldoende gegevens voor herplanverzoek (team en datum nodig)" });
-
-            case VerzoekType.Bevestiging:
-                return JsonConvert.SerializeObject(new { status = "Bevestiging ontvangen", opmerking = "Bevestigingen vereisen handmatige afhandeling door de coördinator" });
-
-            default:
-                return JsonConvert.SerializeObject(new { status = "Niet verwerkt" });
-        }
     }
 
     /// <summary>
